Move the definition of IntrospectionConfig to mytardis_client because conceptually, it is used to hold data we obtain from MyTardis about its own configuration, not for configuring the ingestion behaviour.
