+++ src/config/config.py
@@ -161,31 +161,6 @@ class FilesystemStorageBoxConfig(StorageBoxConfig):
     target_root_dir: Path
 
 
-class IntrospectionConfig(BaseModel):
-    """MyTardis introspection data.
-
-    Pydantic model for MyTardis introspection data. NOTE: this class relies on
-    data from the MyTardis introspection API and therefore can't be instantiated
-    without a request to the specific MyTardis instance.
-
-    Attributes:
-        old_acls : bool
-            the MyTardis instance uses experiment only ACLs if `True`
-        projects_enabled : bool
-            the MyTardis instance uses projects if `True`
-        objects_with_ids : Optional[list[MyTardisObject]]
-    """
-
-    model_config = ConfigDict(use_enum_values=False)
-
-    old_acls: bool
-    projects_enabled: bool
-    identifiers_enabled: bool
-    profiles_enabled: bool
-    objects_with_ids: list[MyTardisObject]
-    objects_with_profiles: list[MyTardisObject]
-
-
 class ConfigFromEnv(BaseSettings):
     """Full MyTardis settings model.
 
@@ -206,12 +181,6 @@ class ConfigFromEnv(BaseSettings):
             instance of Pydantic schema model
         archive: TimeOffsetConfig
             instance of Pydantic time offset model
-
-
-    Properties:
-        mytardis_setup : Optional[IntrospectionConfig] (default: None)
-            instance of Pydantic introspection model either from private
-            attribute or new request
 
     ## Usage
     Requires a .env file in the current working direction:
