+++ tests/test_overseers.py
@@ -14,10 +14,11 @@ from pytest import LogCaptureFixture
 from requests import HTTPError
 from responses import matchers
 
-from src.config.config import ConnectionConfig, IntrospectionConfig
+from src.config.config import ConnectionConfig
 from src.mytardis_client.data_types import URI
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
 from src.mytardis_client.objects import MyTardisObject
+from src.mytardis_client.response_data import IntrospectionConfig
 from src.overseers.overseer import Overseer
 
 logger = logging.getLogger(__name__)
